services/auth-service/requirements.txt
fastapi
uvicorn
pyjwt
pydantic
email-validator
pydantic-settings
sqlalchemy
psycopg2-binary
